+++ src/ResetButton.js
@@ -1,11 +1,16 @@
 function ResetButton (props) {
-// console.log("resetbuttonprops",props)
+console.log("resetbuttonprops",props)
+
+    function resetCounter () {
+        props.setAardbeiCounter(0)
+        props.setBanaanCounter(0)
+        props.setKiwiCounter(0)
+        props.setAppelCounter(0)
+    }
 
     return (
         <button
-            onClick={() => props.setAardbeiCounter(0)}
-        /* onClick props.setbanaan copy paste 3x */
-        >RESET DEZE DING</button>
+            onClick={() => resetCounter()}>RESET DEZE DING</button>
     )
 }
